[REFACTOR]: CompanyInformation SonarQube 이슈 제거

+++ src/main/java/studio/studioeye/domain/company_information/application/CompanyInformationService.java
@@ -47,7 +47,7 @@ public class CompanyInformationService {
         String darkLogoImageUrl = null;
         String sloganImageFileName = null;
         String sloganImageUrl = null;
-        if(lightLogoImage != null && !lightLogoImage.isEmpty()) {
+        if(!lightLogoImage.isEmpty()) {
             ApiResponse<String> updateLogoFileResponse = s3Adapter.uploadFile(lightLogoImage);
             if (updateLogoFileResponse.getStatus().is5xxServerError()) {
                 return ApiResponse.withError(ErrorCode.ERROR_S3_UPDATE_OBJECT);
@@ -55,7 +55,7 @@ public class CompanyInformationService {
             lightLogoImageUrl = updateLogoFileResponse.getData();
             lightLogoImageFileName = lightLogoImage.getOriginalFilename();
         }
-        if(darkLogoImage != null && !darkLogoImage.isEmpty()) {
+        if(!darkLogoImage.isEmpty()) {
             ApiResponse<String> updateLogoFileResponse = s3Adapter.uploadFile(darkLogoImage);
             if (updateLogoFileResponse.getStatus().is5xxServerError()) {
                 return ApiResponse.withError(ErrorCode.ERROR_S3_UPDATE_OBJECT);
@@ -63,7 +63,7 @@ public class CompanyInformationService {
             darkLogoImageUrl = updateLogoFileResponse.getData();
             darkLogoImageFileName = lightLogoImage.getOriginalFilename();
         }
-        if(sloganImage != null && !sloganImage.isEmpty()) {
+        if(!sloganImage.isEmpty()) {
             ApiResponse<String> updateSloganFileResponse = s3Adapter.uploadFile(sloganImage);
 
             if (updateSloganFileResponse.getStatus().is5xxServerError()) {
@@ -77,7 +77,7 @@ public class CompanyInformationService {
         return ApiResponse.ok("회사 정보를 성공적으로 등록하였습니다.", savedCompanyInformation);
     }
 
-    public ApiResponse<CompanyInformation> retrieveAllCampanyInformation() {
+    public ApiResponse<CompanyInformation> retrieveAllCompanyInformation() {
         List<CompanyInformation> companyInformations = companyInformationRepository.findAll();
         if(companyInformations.isEmpty()) {
             return ApiResponse.ok("회사 정보가 존재하지 않습니다.");
@@ -86,7 +86,7 @@ public class CompanyInformationService {
         return ApiResponse.ok("전체 회사 정보를 성공적으로 조회하였습니다.", companyInformation);
     }
 
-    public ApiResponse<String> retrieveCampanyLogoImage(Boolean isLight) {
+    public ApiResponse<String> retrieveCompanyLogoImage(boolean isLight) {
         List<String> logoImageUrls;
         if(isLight) {
             logoImageUrls = companyInformationRepository.findLightLogoImageUrl();
@@ -220,9 +220,9 @@ public class CompanyInformationService {
             return ApiResponse.withError(ErrorCode.COMPANYINFORMATION_IS_EMPTY);
         }
         String fileName = companyInformations.get(0).getLightLogoImageFileName();
-        s3Adapter.deleteFile(fileName);
+        if(fileName != null) s3Adapter.deleteFile(fileName);
         fileName = companyInformations.get(0).getDarkLogoImageFileName();
-        s3Adapter.deleteFile(fileName);
+        if(fileName != null) s3Adapter.deleteFile(fileName);
 
         ApiResponse<String> updateLightLogoFileResponse = s3Adapter.uploadFile(lightLogoImage);
         if (updateLightLogoFileResponse.getStatus().is5xxServerError()) {
@@ -247,7 +247,7 @@ public class CompanyInformationService {
             return ApiResponse.withError(ErrorCode.COMPANYINFORMATION_IS_EMPTY);
         }
         String fileName = companyInformations.get(0).getSloganImageFileName();
-        s3Adapter.deleteFile(fileName);
+        if(fileName != null) s3Adapter.deleteFile(fileName);
 
         ApiResponse<String> updateSloganFileResponse = s3Adapter.uploadFile(sloganImageUrl);
         if (updateSloganFileResponse.getStatus().is5xxServerError()) {
@@ -338,11 +338,11 @@ public class CompanyInformationService {
         }
         for (CompanyInformation companyInformation : companyInformations) {
             String lightLogoImageFileName = companyInformation.getLightLogoImageFileName();
-            s3Adapter.deleteFile(lightLogoImageFileName);
+            if(lightLogoImageFileName != null) s3Adapter.deleteFile(lightLogoImageFileName);
             String darkLogoImageFileName = companyInformation.getDarkLogoImageFileName();
-            s3Adapter.deleteFile(darkLogoImageFileName);
+            if(darkLogoImageFileName != null) s3Adapter.deleteFile(darkLogoImageFileName);
             String sloganFileName = companyInformation.getSloganImageFileName();
-            s3Adapter.deleteFile(sloganFileName);
+            if(sloganFileName != null) s3Adapter.deleteFile(sloganFileName);
             companyInformationRepository.delete(companyInformation);
         }
         return ApiResponse.ok("전체 회사 정보를 성공적으로 삭제했습니다.");
@@ -355,9 +355,9 @@ public class CompanyInformationService {
         }
         for (CompanyInformation companyInformation : companyInformations) {
             String lightLogoImageFileName = companyInformation.getLightLogoImageFileName();
-            s3Adapter.deleteFile(lightLogoImageFileName);
+            if(lightLogoImageFileName != null) s3Adapter.deleteFile(lightLogoImageFileName);
             String darkLogoImageFileName = companyInformation.getDarkLogoImageFileName();
-            s3Adapter.deleteFile(darkLogoImageFileName);
+            if(darkLogoImageFileName != null) s3Adapter.deleteFile(darkLogoImageFileName);
             companyInformation.deleteLogoImage();
             companyInformationRepository.save(companyInformation);
         }
@@ -382,7 +382,7 @@ public class CompanyInformationService {
         }
         for (CompanyInformation companyInformation : companyInformations) {
             String fileName = companyInformation.getSloganImageFileName();
-            s3Adapter.deleteFile(fileName);
+            if(fileName != null) s3Adapter.deleteFile(fileName);
         }
         CompanyInformation companyInformation = companyInformations.get(0);
         companyInformation.deleteCompanyIntroductionInformation();
